Use WORD instead of bare int in truecolor_fill_rect()
Task 4 review flagged a bare `int` for patind in the new
truecolor_fill_rect() function, violating the plan's binding
constraint against bare int/long in new code. Switch it to the
portab.h WORD type; no other bare int/long declarations were
found in the function or file.

+++ vdi/vdi_backend_truecolor.c
@@ -88,7 +88,7 @@ static void truecolor_fill_rect(const VwkAttrib *attr, const Rect *rect)
     WORD x, y, i;
 
     for (y = rect->y1; y <= rect->y2; y++, row += linea_vars.v_lin_wr) {
-        int patind = patmsk & y;   /* starting pattern */
+        WORD patind = patmsk & y;   /* starting pattern */
         UWORD pattern = attr->patptr[patind];
         UWORD *dst = (UWORD *)row;
